% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caller_count.R
\name{add_caller_count}
\alias{add_caller_count}
\title{Adds a column to the dataframe indicating for each phone number
how many times it has already called Silbernetz.
In order to do so, it downloads a caller list from the API and
merges it with the provided dataframe}
\usage{
add_caller_count(numbers)
}
\arguments{
\item{numbers}{Dataframe or tibble with a column called "callers"
that contains the phone numbers}
}
\value{
Returns the provided dataframe/tibble with an additional
column that describes the number of times each phone number has
already called Silbernetz
}
\description{
Adds a column to the dataframe indicating for each phone number
how many times it has already called Silbernetz.
In order to do so, it downloads a caller list from the API and
merges it with the provided dataframe
}
\examples{
\dontrun{
add_caller_count(download_data("Numbers",start_date = "2021-05-01",
end_date = "2021-05-03"))}

}
