% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{add_geodata_to_numbers}
\alias{add_geodata_to_numbers}
\title{Add geographic info to phone numbers}
\usage{
add_geodata_to_numbers(
  numbers,
  geo_cols = c("Ortsnetzname", "Bundesland", "PLZ")
)
}
\arguments{
\item{numbers}{A dataframe of endpoint type "Numbers",
as can be returned by \code{\link[=download_data]{download_data()}}.}

\item{geo_cols}{A character vector that specifies which
geographic information should be added as columns (a subset of "Ortsnetzname",
"Bundesland", "PLZ").}
}
\value{
Returns the input dataframe with new columns added to the right.
}
\description{
Adds columns with information on "Ortsnetzname", "Bundesland" and "PLZ" to
a downloaded dataframe of type "Numbers", as well as a column that
indicates whether a call came from landline.
}
\details{
Note that phone codes ("Ortsnetzkennzahl") can overlap
with multiple post codes ("PLZ") and vice versa. Therefore, post codes
are encoded as character vectors.

Geographic data has been compiled from multiple sources,
all of which have been accessed between July and August 2021:
\itemize{
\item Bundesnetzagentur
\item http://www.suche-postleitzahl.org/downloads
\item http://www.xn--datenbrse-57a.net/item/Postleitzahlen-Datenbank_Deutschland
\item http://wogibtes.info
}
}
\examples{
\dontrun{
add_geodata_to_numbers(download_data("Numbers",
start_date = "2021-05-31", end_date = "2021-05-31"))
}

}
