% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_current_data_file.R
\name{create_current_data}
\alias{create_current_data}
\title{Function to create current data file
This file is supposed to be called once per installation}
\usage{
create_current_data(end_date, write_path = "../data/raw/current_data.csv")
}
\arguments{
\item{end_date:}{Current date, formate "YYYY-MM-DD"}

\item{write_path:}{Where to write the resulting csv, should correspond the "file" variable in shinyapp.R}
}
\value{
None, only side effect of writing the file
}
\description{
Function to create current data file
This file is supposed to be called once per installation
}
