% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_herkunft_table.R
\name{create_herkunft_table}
\alias{create_herkunft_table}
\title{Function to build Elke's Herkunft Table which summarises
the number of calls per Bundesland across weeks}
\usage{
create_herkunft_table(df, start_date, end_date)
}
\arguments{
\item{df}{Dataframe/Tibble containing the input data}

\item{start_date}{Start of the reporting period}

\item{end_date}{End of the reporting period}
}
\value{
Returns table
}
\description{
Function to build Elke's Herkunft Table which summarises
the number of calls per Bundesland across weeks
}
