% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_provider_table.R
\name{create_provider_table}
\alias{create_provider_table}
\title{Function to build Elke's Provider Table which summarises
the number of calls per week or month}
\usage{
create_provider_table(df, unit, period_start, period_stop)
}
\arguments{
\item{df}{Dataframe/Tibble containing the input data}

\item{unit}{Aggregate to either weeks ("week"), months ("month") or days ("day") or years ("year")}

\item{period_start}{Start of the reporting period}

\item{period_stop}{End of the reporting perios}
}
\value{
Returns Elke's provider table
}
\description{
Function to build Elke's Provider Table which summarises
the number of calls per week or month
}
