% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download Silternetz raw data}
\usage{
download_data(
  endpoint,
  start_date = NULL,
  end_date = NULL,
  verbose = TRUE,
  create_firstcall_column = TRUE
)
}
\arguments{
\item{endpoint}{Specifies which data to download. Has to be one of
one of Dest, Dest_Count, Numbers, Numbers_Count, Callerlists.}

\item{start_date}{First date of the period of interest. Format: YYYY-MM-DD.}

\item{end_date}{Last date of the period of interest. Format: YYYY-MM-DD.}

\item{verbose}{Boolean if verbose mode is turned on}

\item{create_firstcall_column}{whether a column should be created that
shows that a call was a first call by a caller. When the function is
called from within the Silbernetz app, this should be \code{FALSE} because the app
does its own handling of the first caller column (saving time).}
}
\value{
Downloaded dataframe.
}
\description{
Provides a connection to the INOPLA-API and allows to download the data from
specified periods.
}
\details{
The information in ouuput column "firstcall" refers to whether a call has been the first
one of the respective caller among those in the time frame that is being downloaded.
}
\examples{
\dontrun{
download_data("Dest", start_date = "2021-05-01", end_date = "2021-05-31")}

}
