% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_laender.R
\name{heatmap_laender}
\alias{heatmap_laender}
\title{Bundeslaender heatmap}
\usage{
heatmap_laender(
  numbers_geo,
  variable,
  start_date = NULL,
  end_date = NULL,
  filter_success = FALSE,
  show_zeros = FALSE
)
}
\arguments{
\item{numbers_geo}{A dataframe of endpoint type "Numbers" with geographic data
added, as returned by \code{\link[=add_geodata_to_numbers]{add_geodata_to_numbers()}}
as can be returned by \code{\link[=download_data]{download_data()}}.}

\item{variable}{Which variable should be counted? Either "Anrufe", "Anrufer" oder "Erstanrufer".}

\item{start_date}{Start date of time span to be plotted (if not specified,
minimum date in \code{numbers_geo} is used).}

\item{end_date}{End date of time span to be plotted (if not specified,
maximum date in \code{numbers_geo} is used).}

\item{filter_success}{If only successfull calls ("Verbunden") should
be taken into account.}

\item{show_zeros}{If areas with zero count should be included in the color scale
as numeric zeros. Default alternative is to show them as grey areas without any label.}
}
\value{
Returns a ggplot.
}
\description{
Creates a heatmap for frequency of calls or callers by Bundesland for
a specified time period.
}
\examples{
\dontrun{
heatmap_laender(add_geodata_to_numbers(download_data("Numbers",
start_date = "2021-05-30", end_date = "2021-05-31")),
variable = "Anrufer")
}

}
