% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calls_over_time.R
\name{plot_CallsOverTime}
\alias{plot_CallsOverTime}
\title{Plot number of calls over time}
\usage{
plot_CallsOverTime(
  data,
  show_success,
  show_firstcall,
  start_date,
  end_date,
  nr_breaks,
  state = "Alle",
  y_lim_zero = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{show_success}{Boolean whether only successfull calls should be shown}

\item{show_firstcall}{Boolean whether only first calls should be shown}

\item{start_date}{start date, format "YYYY-MM-DD"}

\item{end_date}{End date, format "YYYY-MM-DD"}

\item{y_lim_zero}{Whether the lower limit of the y axis should be zero}
}
\value{
plot
}
\description{
Plot number of calls over time
}
