% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_redundant_cols.R
\name{remove_redundant_cols}
\alias{remove_redundant_cols}
\title{Removes columns that are provided by the API but not needed by us}
\usage{
remove_redundant_cols(
  data,
  red_cols = c("service", "service_caption", "tarif", "modul_name")
)
}
\arguments{
\item{data}{data set with data}

\item{red_cols}{Columns to remove, defaults to "service", "service_caption",
"tarif" and "modul_name".}
}
\value{
Data with columns removed
}
\description{
Removes columns that are provided by the API but not needed by us
}
